// src/shared/models/accion-correctiva-analisis-hijo-5ws.ts
export interface AccionCorrectivaAnalisisHijo5wsModel {
    id?: number;
    id_accion_correctiva_analisis?: number;
    quien?: string;
    que?: string;
    cuando?: string;
    donde?: string;
    como_ocurre?: string;
    como_resolverlo?: string;
    id_usuario?: number;
    fecha_creacion?: number;
}